new color configuration dialog.
The goal of this dialog is to be simpler to use and fit a small
screen, however provide more useful features to the user.

This was done by having one single multi-select ilist with color
previews on the left (just text now, but supports solid colors as
well) and a full-featured preview on the editor box.

The list supports multi-select, so one can enable, disable or set
colors on multiple items at the same time, avoiding the old
enable/disable radios that did not make much sense.  The frame label
should be updated based on the items one select (if they have the same
colors or are "mixed", if they are all unset).

The list also uses the new feature to show an "end" icon to notify the
item is customized or not.

TODO: review supported/known color classes, it is as simple as
      changing the description array.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@46928 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33