Issue when you cut (or copy) and then paste a part Bug: When you cut (or copy) a part and then paste it, the state field was not set right, it was set like @ani@0.30 0.00 for example.
Resolution: Save the selected state when cut/copy then select it again
when paste.
The function obj.state_selected_set(*self["state"]) was used in function
apply_to(self, obj) in Part class. This seems not to be a problem
because the apply_to function must set in the obj exactly what is in
the Part object.

By: Helen Fornazier <helen.fornazier@profusion.mobi>



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@49033 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33