Do not activate DPMS, Screen Saver or Locker if there is a fullscreen window.
I guess there is nothing more annoying than having a fullscreen window
to watch a quick movie, read some text or webpage or whatever and have
the DPMS to show, or the screensaver to be activated or your screen to
be locked.

This patch uses the new border signal and counts, if there is at least
one fullscreen border, then avoid doing these annoyances.

I'm not accounting for any border parameter/property, like if it is in
a visible screen or desktop. I hope this is fine, otherwise let me know.

I also asked and common sense at IRC (at least to me and englebass) is
that there is no need to configure such behavior, so there is no
configure option for it.

I'm also thinking about maximized windows to have a configurable
behavior to match this, but I'm not sure.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@49262 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33