Fix current event pointer when walking filters
When walking the filters' list, using event_current might screw a
recursive main loop that had to process filters *and* events on the
first iteration. Thus, use a new pointer to mark the current event being
treated when walking filters' list.

Tests for this one is difficult. I'll try to think about a test later.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@49435 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33