Although evas_debug is flagged with "debug" in the name, messages are critical.
The evas_debug functions are just called from MAGIC_DEBUG errors, that
are quite critical, so messages should come out by default, and do
that loud with a CRITICAL warning.

Raster, cedric and others: now that we have Eina as a common
dependency between Evas and Ecore, maybe it is worth to unify the
debug system on top of Eina logging? It already provides "abort on
critical", runtime configurable log levels. I'd say the same about the
magic and safety checks themselves (eina_magic and eina_safety).




git-svn-id: http://svn.enlightenment.org/svn/e/trunk@49643 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33