Modify protocols to take a reference to the Emote_Protocol structure as a parameter to protocol_init.  This will allow them to guard against events destined for other protocols.
Add emote_protocol_list which will return a list of strings representing
the protocols.  The name can be passed to emote_protocol_load to acutally
load the protocol.

Modify app so it gets the protocol list and sends a simple event to each
protocol available.  All protocols will respond with an event when this
event is received.

There does seem to be an issue with repeat messages, but I'm tired and
can't figure it out so if someone could look into that it would be good.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@51566 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33