...even on unload. Fixed a bug when the process couldn't be spawn and the module get's unloaded.
I vote for removing the module or finishing it, what it currently does can be done with a simple .desktop file which is started on e17 start.

In the end this module looks like a quick hack: There should be a messages if openoffice couldn't be found, there should be some code to find the executable via the .desktop file instead of hardcoding a executable, etc.
Also the prestart process quiets after running openoffice, so there should be some code to check periodly for the process...


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@51940 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33