Don't cast Eina_Inlist directly to Elm_Genlist_Item.
The correct way is to use EINA_INLIST_CONTAINER_GET() that allows the
Eina_Inlist structure to be moved around in the structure. Right now
it did work fine, but I want to change the Elm_Genlist_Item later and
this change will be required and it is better to come in a separate
patch.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@52112 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33