Fix preferences x command line option handling.
As it is hard to define what to do in such case, the command line
option will change the preferences.

Some will argue this is not the right thing to do, but then the
preferences screens would be incorrect. In order to simplify the
semantics I now declare this "the right thing to do".



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@52231 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33