Begins highlight on focused object.
There's still work to do here, particularly in the theme, but it has
something nice and fun to see the code working.
The idea behind this:
Window tracks focused object and sends the highlight object(s) to it. These
are simple edje objects, one on top, one below the focused widget for nice
effects. Widgets can choose to ignore the highlight and this will be sent to
the parent object, if it doesn't ignore it as well.
About the bottom object, it doesn't work now. For the most part, focused
widget will always be a member of some smart object, so stacking won't work
and the desired effect is nowhere to be seen. This will be worked out later.
To be done now:
 - Let the theme for a widget define its own highlight, disabling if needed
 the standard one for those objects.
 - Needed base in code to allow animations when switching focus. All in theme.
 - Properly test all widgets and fix some things that will most likely work
 in weird ways, given the nature of Evas/Edje and how Elementary makes use
 of them.
 - Forgot the rest, stay tuned, test, report, give ideas, plant a tree.

Work started by glima, continued with some refactors by me when he
decided he needed vacations.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@52524 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33