ephoto core simplification - part 1 (UNFINISHED!!!!)
I'm rewriting parts of ephoto to simplify it, one of the things is to
avoid lots of nesting of layouts to get a simple preview
image.

So far just the thumb directory is done to the point it shows the
contents of a directory, it will do nothing on selection and
clicks. There is no toolbar either, as this is going to be replaced
later on.

EFL bugs so far:
 * eio_file_cancel() is not playing nice with ecore_pipe, with errors like: "ecore_pipe.c:517 _ecore_pipe_read() An unhandled error (ret: -1 errno: 9)occurred while reading from the pipe the length"

 * evas/edje map is broken WRT clipping! if I have map { on: 1;
   rotation.z: 10;} and this part is clipped to some other part, the
   clipper will be ignored!

Sorry, but I couldn't finish it now, trying to do it later today and
tomorrow. I still need to add the elm_genlist_item_prepend(),
elm_genlist_item_insert_before() and elm_genlist_item_insert_after()
to get the sorted insert to work.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@53237 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33