[Elementary] Implement priority in Elm_Toolbar.
Toolbars can now be scrollable, non-scrollable, or hide icons selectively
according to a priority.  This changed the API slightly, thus

   elm_toolbar_scrollable_set(obj, 1)

Should be

   elm_toolbar_mode_expand_set(obj, ELM_TOOLBAR_EXPAND_SCROLL)

Values for non-scrollable and scrollable remain the same, so if constants
aren't being used, just the function name should be changed.

The new mode, ELM_TOOLBAR_EXPAND_MENU, will create a menu of items when the
space isn't large enough to hold all the items.  Which items will be kept on
the toolbar or which items will be added to the menu is controlled by the
priority of each item: higher ones will be kept if possible, lower ones will
be tratored on the first opportunity.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@53406 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33