Add tests to elementary_test to exercise toolbar's shrink modes.
Now that toolbar has 3 shrink modes I've added one test for each one. The first
one shows the menu mode which is responsible for adding a More button when the
toolbar size is not large enough to hold all items. The second one is scroll
mode in which the items can be scrolled when the length is not enough. The last
one, none mode, does nothing.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@53718 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33