coccinelle: silence autogen
In general, when running spatchall -a we are not interested in the output of
autogen.sh. Worse, in the case the script is run in report mode, the output
would be cluttered by useless output. So, let autogen.sh to be silent.

For example, run spatchall.sh as below and you will see several places in
which realloc is used passing the same pointer as parameter and return value,
possibly losing the pointer in case of failure.

	./spatchall.sh -a realloc-spot-lost-pointer.cocci 2>/tmp/stderr \
		| tee /tmp/stdout



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@53854 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33