From: ChunEon Park <chuneon.park@samsung.com> Subject: [PATCH] elc_ctxpopup
Here code is for the ctxpopup. 

Since the functionalities of ctxpopup are somewhat different, 
I'm trying to overwrite the ctxpopup code entirely first. 

The main reason is, 
Original ctxpopup could determine it's size and position by itself.
Of course, upstreamed version supports this. But it can not be reduced
it's
size
And it supposes the arrow position to center of box edje. 

Basically, 
This version calculates it's box size first then find available space
from
the it's position. 
The origin position of the box is not expected, it finds the available
positions. 
But if there is not enough space to be shown, then it finds a best
direction
to be shown again.
In this case, it reduces it's size also as to be shown entirely. 
Since the ctxpopup supports scroller, the reduced box could show the all
items also. 
(The point is ctxpopup should know the size of total items...)

I'm not sure the upstreamed version could cover the above scenario, 
If then, I need to modify it first.

(fixed some minor formatting issue too)



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@56021 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33