From: Mike McCormack <mj.mccormack@samsung.com>
ecore_config code may end up closing stdin, due to assuming 0 is a valid fd.

I know this code is deprecated, and it's being kept around "for reference".

Either:
A) this bug should still be fixed so bad code is not copied elsewhere

OR

B) ecore_config should be deleted so bad code is not copied elsewhere

thanks,

Mike


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@56275 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33