From: 김재환 <jae.hwan.kim@samsung.com> Subject: [E-devel] [Patch] els_scroller - Cannot click even if it reach to end
I have another issue in scroller when the bounce is off.
Scroller cannot be clicked even if it reach to edge and animation is
end.
The reason is why the momentum animation is running even if it reach
end and bounce animation is off.
So I change the code that if it reach the edge in case of bounce-off,
the momentum animation is stoped.
The following is the patch code.




git-svn-id: http://svn.enlightenment.org/svn/e/trunk@56580 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33