From: Jae Hwan Kim <jae.hwan.kim@samsung.com> Subject: [E-devel]  [Patch] elm_smart_scroller_child_region_show_internal
In the els_scroller of the elementary, the function
"elm_smart_scroller_child_region_show_internal"
have a problem. If the parameters "x,y,w,h" have the wrong value like
"elm_scroller_region_show(scroller,
-300, -300, 900, 900);", the scroller move the position temporarily.
The scroller seem to jump.
So if the value of the parameter is not valid, it should be changed to
valid value.
The work is applied already in the function
"elm_smart_scroller_region_bring_in".
It should be applied in the function
"elm_smart_scroller_child_region_show_internal" too.
I changed the code like bring_in function.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@57666 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33