Ecore: Ecore_X: Xlib: Fix memory leak.
When handling xdnd_enter event(s), if we do not support the dnd target
version, then we issue a warning and return from the handling
function. If we are going to return (and not issue the ecore_x_event),
then free the allocated memory of the ecore_x_event_xdnd_enter
structure that we previously allocated.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@58337 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33