From: Ulrich Eckhardt <doomster@knuut.de> Subject: [E-devel] XRender engine causes ecore build failure
while building ecore. The problem is that this engine was removed from evas 
but not yet completely from ecore. I was on IRC with Vincent Torri (vtorri) 
and Daniel Juyung Seo (SeoZ) and the consensus was to remove the code for the 
XRender engines, both the Xlib and XCB versions.

There is a switch over the different engine types, where there are still a few 
places left where XRender is handled, grep for "xrender" or "XRENDER" and you 
will find them. The question is whether to just return NULL in order to signal 
that this engine is not supported or to remove the whole thing. The latter 
could break binary compatibility, therefore I left those stubs in.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@60502 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33