evas/autotools - fix build when no dependencies for examples are found.
Some examples need ecore, ecore-evas and edje to demonstrate its
functionality. If --enable-build-examples is used, these dependencies are
required. This commit disables the examples build if they are not found
in the system.

TODO: improve efl_examples.m4 and use it in evas build too - I'll work on
that soon.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@60609 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33