ecore: do count threads only in the main loop.
NOTE: for an unknow reason I always get the wrong number
of threads when doing the computation from the thread.
Even if I use volatile and mutex. So to avoid that move
that stuff in the main loop. It increase the complexity
of the code, but at least it work.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@60767 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33