From: Christian Apeltauer <c.apeltauer@gmx.de> Subject: [E-devel] Proposed patch to elsa
ear developers,
 since some weeks now I have trouble running elsa: The X server starts,
 but elsa_client never comes up. I was able to hunt down the
 error. Currently I am using rev 60908.
  In elsa/src/daemin/elsa_server.c lines 135-142:
  
     dirs = efreet_data_dirs_get();
     EINA_LIST_FREE(dirs, path)
       {
          snprintf(buf, sizeof(buf), "%s/xsessions", path);
          _elsa_server_scan_desktops(buf);
          eina_stringshare_del(path);
       }
     efreet_shutdown();

The data structure dirs points to is freed, but the static variable
xdg_data_dirs (defined in efreet/src/lib/efreet_base.c:48) still points
to it. efreet_shutdown() leads to efreet/src/lib/efreet_base.c:87 

    IF_FREE_LIST(xdg_data_dirs,eina_stringshare_del);
    
    The macro IF_FREE_LIST checks whether xdg_data_dirs is a NULL pointer
(which it is not) but not whether the Eina_List pointed to is already
freed and tries to free it again leading to a hang up of elsa. So I
suggest the following patch:



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@61035 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33