emage: please welcome emage. It's now something like a file browser.
NOTE: it started as a benchmark for emotion, so browse into a directory
full of video and see them in a live thumbnail. You can have them all
playing at the same time by passing -c 1 to emage command line. I recommand
in that case to use OpenGL backend for Ecore_Evas and Gstreamer for
Emotion backend (use the environment variable : EMOTION_ENGINE=gstreamer
ECORE_EVAS_ENGINE=opengl_x11).

NOTE2: as it was easy to add other file format, I hacked in a image
viewer as it is now a requirement for every one to write his own image
viewer. But take that more as an example on how to add another file
format, than anything else.

NOTE3: it's free software, I have no clear goal with that application
except to benchmark emotion, so if you want to hack in feel free to
do so, you are welcome !



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@61054 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33