Evas font+textblock: Stopped using FcNameParse, parse ourselves.
FcNameParse is pretty bad, not only that it's pretty bad, it doesn't
work well in some cases. Also this makes our lives a lot easier since
we want to allow overriding style/weight/width and etc from markup.
We now parse the "font" markup ourselves, populating a font description
structure that we created, and then override values in it as needed.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@61936 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33