Theme detourious: Don't embed the font in theme.
There's no font fallback from edje -> fontconfig so non-latin languages
won't work in e if we do. That's a flaw in evas, I guess, but until it's
solved, we can't do it. Especially not with a candidate to be default
theme.

The change tries searching for the Switzera font in the system, and if
it exists, it takes it, otherwise, it falls back to another font
according to fontconfig rules.

There's an issue with this change, because the font= valuse should be
enclosed inside quotes, i.e font='FONT' so things like:
"Switzera:style=bold italic" will work.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@62652 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33