enjoy: improved plugin infrastructure.
Plugins can now be disabled and enabled (at least in infrastructure
terms, no UI or code to filter out enabled plugins).

Two new signals will be emitted to main loop:
  - ENJOY_EVENT_STARTED
  - ENJOY_EVENT_QUIT

Plugins will be enabled on start, disabled on quit. Quit will also
preserve the main loop until everything is done using it. It is
required for stuff like FSO that needs to talk to DBus to re-enable
the CPU policy.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@63163 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33