eina: add eina_inlist_sorted_state_insert and helper.
Note: this function help keep a jump table so we reduce
the need to walk over the complete list to insert one
element. It's of course doesn't make it an O(log(n)) in
access time, but it increase it's cost more slowly.
   With 10000 items, you can count around 50 pointers
dereferencing and with with 50000 items around 200 pointers
dereferencing.
   Of course the comparison stay in O(log(n)).


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@63213 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33