THEMES/detorious: try a simpler pager.
Couple of people reported that they found the pager a bit strange, I
also felt the same, then I'm trying a new one that is more clean in
the design.

Pager is usually quite small and holds windows, possibly lots of them
(xterms side by side), in this case all those details were cluttering
the results.

Replaced the sunken border with a single line at left and right, the
focused background is a plain rectangle with the focused color. This
should look fine in both inset and plain styles.

Windows have no detail. Unfocused are semi-transparent, focused is
opaque. Just focused window have icon, and the icon should be at least
16x16.

If you dislike it, feel free to revert.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@65128 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33