elementary: Add multibuttonentry widget
---------------------------------------------------

Dear Developers,

I added a new widget called “Multibuttonentry" which could be used as contact button in App
This widget allows a user to insert a text button.
text button is inserted by pressing the "Enter key"
If there is no space in the current row, the new button is entered in the next row. If the button is clicked, it will become focused.
And focused button can be removed by pressing the "backspace" key.
when items are added over 1 lines, if Multibuttonentry lost focus, it becase shrink mode (made it as 1 line)

Please review it's APIs and functionalities.

Thank you,
Goun

---------------------------------------------------

Widget was originally written by Hyunsil Park <hyunsil.park@samsung.com>.
It has been contributed to by the following people:
 * Jihoon Kim <jihoon48.kim@samsung.com>
 * WooHyun Jung <wh0705.jung@samsung.com>
 * deasung.kim <deasung.kim@samsung.com>
 * Myungjae Lee <mjae.lee@samsung.com>
 * Goun Lee <gouni.lee@samsung.com>
 * Mike McCormack <mj.mccormack@samsung.com>
 * Daniel Juyung Seo <juyung.seo@samsung.com>
 * Seunggyun Kim <sgyun.kim@samsung.com>
 * sanghoon80.cho <sanghoon80.cho@samsung.com>

Signed-off-by: 이고은 <gouni.lee@samsung.com>
Signed-off-by: Mike McCormack <mj.mccormack@samsung.com>

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@65548 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33