Bug fix - about type change in group inherit. Lately, raster removed the code about the prohibition of type-change in group inherit. But about the "part" of different type, the data structure of the their "description" is different. So if the type is changed, it have to be reallocated. Current, it is not.
At first, we have to remove the lookups. If we don't, when lookup module
executes, the memory
may be broken. So I removed all lookups for reallocated description before
it is reallocated.
And I changed all description of the "part" is reallocated when the type is
changed.
The attribute of the "part" is remained. Just it reallocated the part of
**_Spec_**.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@66242 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33