From: Bluezery <ohpowel@gmail.com> Subject: [E-devel] [Patch][elm_map] Smooth stepping of wheel & pinch gesture
When elm_map doing zooming by using wheel gesture, scroller is resized
if it exceeds certain limit.
But current limit value is 2.0 & 1.0, actually those should be below
2.0 and above 1.0.
The reason is that resizing means exactly 2.0 or 1.0. So if it is
above 1.99x,  it can be 2.0 and can be resized.
In addition, pinch.level & pinch.diff values should not be 1.0 or 2.0
for more accuracy. It should be half of previous pinch level or two
times of that.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@66658 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33