elev8: Add 'contents' parameter to CElmLayout
This allows one to instantiate Elev8 objects and swallow them in the layout.
For example, the following declaration:

   {
      type: 'layout',
      ...,
      contents: {
         'file-list': {
            type: 'genlist',
            ...
         },
         'dir-list': {
            type: 'genlist',
            ...
         }
      }
   }

Will instantiate two genlists and swallow them in their respective places in
the layout.  ``contents'' has been choosen instead of ``elements'' because
that's what the Elementary layout calls this inside
(elm_object_part_*content*_set()).



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@68080 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33