evas: remove software SDL engine, use buffer engine directly.
NOTE: I would like to do the same with software SDL 16bits engine.
But as we don't have a buffer_16 backend, I am likely to just remove
it and use buffer conversion code to match a 16bits target. This
will come with a performance impact, that will make it useless. So
I am just tempted to completly remove it.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@68352 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33