From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [Patch][elm_list] bug fix: crash occurs when container which contains the elm_list is deleted on the "elm_list item clicked" callback
I have resolved the elm_list issue(goto ELM_LIST_CRASH_ISSUE_MAIL below, I
sent mail but I cannot find the previous mail -_-;).
It would be related with double free issue. In the "sub-obejct-del"
callback which name is _sub_del() sets item->icon to NULL.
And in the EVAS_CALLBACK_MOUSE_UP callback funtion which name is
_mouse_up() calls evas_object_unref();
It would try to free the item->icon but the item->icon is already set to
NULL. So.. the crash would occur..



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@69555 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33