Ecore_Wayland: Add an API function which calls wl_display_iterate(readable).
Fix ecore_wl_window surface attach of the buffer for shm and egl types.

NB: This ensures that both efl/elm clients work under the Wayland
Compositor AND the E17 Compositor.

NB: The issue here is that clients (when run under the Wayland
compositor) need to make sure 'iterate' has
been called before creating surfaces so that the wl_compositor is
setup. However, the actual E17 Compositor will hang if 'iterate' is called
to early.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@69707 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33