Background object revised, now match the C api.
CALL FOR HELP:
Ok we need to do this work for each elm widget, it's a really 
annoing work to do, but it's also very easy.

Someone want to help me in this revision work? It will take months for me to do alone

If you are intrested I'm using the c_elementary.pxd file as index,
I will put a (DONE) on widget that are...well..done :)

davemds



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@70937 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33