From: Michal Pakula vel Rutka <m.pakula@samsung.com> Subject: [E-devel] [patch] allow copy and paste operations for non-widget evas objects
I am currently implementing a clipboard for WebKit-EFL. When using
elm_cnp_selection_get and set function an evas object passed has to be
a widget. Currently in WebKit-EFL we are using an view which is an
Evas_Object, but not a widget. Sample WebKit applications like
EWebLauncher and tests applications does not use elementary so I
cannot obtain parent widget for a view object and finally we cannot
perform copy and paste operations.
Widgets are used in elm_cnp_selection functions to obtain
Ecore_X_Window. My patch leaves current way of obtaining it and adds
an another in case when the standard one fails. When Evas_Object is
not a widget I am trying to get an Ecore_X_Window using ecore.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@71381 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33