From: Bluezery <ohpowel@gmail.com> Subject: [E-devel] [Patch][elm_map] Add New "loaded" signal
I want to add new "loaded" signal to elementary map. This signal
notify users when all map tiles are loaded. Timer is used for this signal.
Currently "tile,loaded" can be used for each separated tile.  So the
users are hard to know whether all tiles are loaded or not.
Please review this patch.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@71406 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33