elev8: Connect to EVAS_CALLBACK_FREE event.
This callback is called after the Evas_Object has been freed from the EFL
side.  When this event is received, a debug message is printed and the
pointer to said Evas_Object is nullified.

Hopefully magic checks inside Elementary will print the expected object type
when it finds a null pointer, giving a hint of where to look for possible
errors.

This shouldn't happen in normal circumstances, but it is a good measure to
avoid crashes and hunt for bugs.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@72620 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33