welcome econnman application.
This application attempts to cope with newest connman API and provide
every feature it supports.

It's written as an application instead of E17 module to make use of
Elementary. At this point I'm trying to make something usable fast and
then python-elementary is used (it may be converted to C at a later
point).

Just use a recent elementary from svn to run, as elm_segment_control
was badly broken and will segfault if used! Be warned.

python-elementary is not a compile time dependency and thus not
checked by configure, but it's runtime so be sure to have it before
running!

TODO:
 - IPv6
 - Proxy
 - Technology and Service icons (IOW: it looks ugly now)



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@73121 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33