e_dbus: Use idler instead of idler_enterer
Instead of using an idle_enterer to process D-Bus messages, use an
idler that does not require creating a dummy event and wakeup the main
loop on every interation.

Remove a callback function to wake up the main loop when there is a message
in queue.


Patch by: José Roberto de Souza  <zehortigoza@profusion.mobi>



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@73371 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33